[REEF-146] Changing .Net project structure for Reef tests, reef client and Reef all
Move project folders under cs
Rename projects to follow name convention
Update namespace for projects
Update csproj file for the references
Rename reef test folders and update namespace inside reef tests

JIRA: Reef-146. (https://issues.apache.org/jira/browse/REEF-146)
Author: Julia Wang  Email: jwang98052@yahoo.com
