+++ lang/cs/Org.Apache.REEF.Tests/Utility/TestExceptions.cs
@@ -1,0 +1,49 @@
+﻿/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *   http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied.  See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+
+using System;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Org.Apache.REEF.Utilities.Diagnostics;
+using Org.Apache.REEF.Utilities.Logging;
+
+namespace Org.Apache.REEF.Tests.Utility
+{
+    [TestClass]
+    public class TestExceptions
+    {
+        private static readonly Logger LOGGER = Logger.GetLogger(typeof(TestExceptions));
+
+        [TestMethod]
+        public void TestThrowCaught()
+        {
+            string msg = null;
+            try
+            {
+                Exceptions.Throw(new ApplicationException("test"), LOGGER);
+                msg = "not supposed to reach here";
+                Assert.Fail(msg);
+            }
+            catch (ApplicationException e)
+            {
+                Exceptions.Caught(e, Level.Info, LOGGER);
+            }
+            Assert.IsNull(msg);
+        }
+    }
+}
